Next >> Developer Interface

nextstream retrieves the currently stored L'Ecuyer-CMRG RNG stream for the specified compute profile and advances it to the next stream.

nextget retrieves the specified item from the specified compute profile.

nextcode translates integer exit codes returned by daemon().

Usage

nextstream(.compute = "default")

nextget(x, .compute = "default")

nextcode(xc)

Arguments

.compute

[default NULL] character value for the compute profile to use (each has its own independent set of daemons), or NULL to use the 'default' profile.

x

character value of item to retrieve. One of "n" (number of dispatcher daemons), "pid" (dispatcher process ID), "dispatcher" (the URL to connect to dispatcher from the host) "url" (the URL to connect to dispatcher from daemons) or "tls" (the stored client TLS configuration for use by daemons).

xc

integer return value of daemon().

Value

For nextstream: a length 7 integer vector, as given by .Random.seed when the L'Ecuyer-CMRG RNG is in use (may be passed directly to the rs argument of daemon()), or else NULL if a stream has not yet been created.

For nextget: the requested item, or else NULL if not present.

For nextcode: character string.

Details

These functions are exported for use by packages extending mirai with alternative launchers of daemon() processes.

For nextstream: This function should be called for its return value when required. The function also has the side effect of automatically advancing the stream stored within the compute profile. This ensures that the next recursive stream is returned when the function is called again.

Examples

if (FALSE) { # interactive()
daemons(1L)
nextstream()
nextstream()

nextget("pid")
nextget("url")

daemons(0)
}
nextcode(0L)
#> [1] "0 | Daemon connection terminated"
nextcode(1L)
#> [1] "1 | Daemon idletime limit reached"