SSH Arguments Constructor

Return value may be supplied directly to the 'args' argument of launch_remote or make_cluster.

ssh_args(names, port = 22, timeout = 5, tunnel = FALSE)

Arguments

names

a character vector of hostnames or IP addresses of the remote machines on which to launch daemons (nodes), e.g. c('10.75.37.90', 'nodename').

port

[default 22] numeric port number on which to connect.

timeout

[default 5] maximum time allowed for connection setup in seconds.

tunnel

[default FALSE] logical value whether to use SSH reverse tunnelling. If TRUE, a tunnel is created between the same local and remote port specified as part of 'url' of the function where this is being used as an argument. 'url' in this case must be either 'localhost' or '127.0.0.1'. This is as the host listens at 'url' on the local machine, whilst the nodes dial into the same 'url' on the remote machine, and the SSH tunnel connects both ends.

Value

A list of character vectors the same length as 'names'.

Details

By default, uses the SSH flags -f to background the SSH process and -T to not require a tty for efficiency.

For simplicity, this SSH tunnelling implementation uses the same port on both the side of the host and that of the corresponding node.

Note

This function only returns the relevant SSH arguments; 'command' must also be specified as 'ssh' or the path to the SSH application.

SSH Direct Connection

The simplest use of SSH is to execute the daemon launch command on a remote machine, for it to dial back to the host / dispatcher URL. The relevant port on the host must be open to inbound connections from the remote machine.

The following shows an example for make_cluster, and is equally valid for the same arguments in launch_remote.

make_cluster( # host URL accessible to nodes url = "tls+tcp://10.75.37.40:5555", # launch nodes using SSH command = "ssh", # node IP / hostnames to connect to with custom port number args = ssh_args( names = c("10.75.37.90", "10.75.37.91"), port = 222, timeout = 1 ) )

SSH Tunnelling

Use of SSH tunnelling provides a convenient way to launch remote nodes without requiring the remote machine to be able to access the host. Often firewall configurations or security policies may prevent opening a port to accept outside connections.

In these cases SSH tunnelling offers a solution by creating a tunnel once the initial SSH connection is made. The tunnel is created between localhost:port or equivalently 127.0.0.1:port on each machine. This means that the host should listen to localhost:port and the nodes should each dial into localhost:port on their own respective machines.

The following is an example for make_cluster, and is equally valid for the same arguments in launch_remote.

make_cluster( # use 'localhost' or '127.0.0.1' with a port available on all machines url = "tcp://localhost:5555", # launch nodes using SSH command = "ssh", # launch 2 nodes on the remote machine with SSH tunnelling args = ssh_args( names = c("10.75.37.90", "10.75.37.90"), timeout = 1, tunnel = TRUE ) )

Examples

ssh_args(names = c("10.75.37.90", "nodename"), port = 222, timeout = 10)
#> [[1]]
#> [1] "-o ConnectTimeout=10 -fTp" "222"                      
#> [3] "10.75.37.90"               "."                        
#> 
#> [[2]]
#> [1] "-o ConnectTimeout=10 -fTp" "222"                      
#> [3] "nodename"                  "."                        
#>