mirai Dispatcher

Implements a dispatcher for tasks from a client to multiple servers for processing, using a FIFO scheduling rule, queuing tasks as required.

dispatcher(
  client,
  url = NULL,
  n = NULL,
  asyncdial = FALSE,
  token = FALSE,
  lock = FALSE,
  ...,
  monitor = NULL
)

Arguments

client

the client URL to dial as a character string (where tasks are sent from), including the port to connect to and (optionally) a path for websocket URLs e.g. 'tcp://192.168.0.2:5555' or 'ws://192.168.0.2:5555/path'.

url

(optional) the URL or range of URLs the dispatcher should listen at as a character vector, including the port to connect to and (optionally) a path for websocket URLs e.g. 'tcp://192.168.0.2:5555' or 'ws://192.168.0.2:5555/path'. Tasks are sent to servers dialled into these URLs. If not supplied, 'n' URLs accessible from the same computer will be assigned automatically.

n

(optional) if specified, the integer number of servers to listen for. Otherwise 'n' will be inferred from the number of URLs supplied as '...'. Where a single URL is supplied and 'n' > 1, 'n' unique URLs will be automatically assigned for servers to dial into.

asyncdial

[default FALSE] whether to perform dials asynchronously. The default FALSE will error if a connection is not immediately possible (e.g. daemons has yet to be called on the client, or the specified port is not open etc.). Specifying TRUE continues retrying (indefinitely) if not immediately successful, which is more resilient but can mask potential connection issues.

token

[default FALSE] if TRUE, appends a unique 40-character token to each URL path the dispatcher listens at (not applicable for TCP URLs which do not accept a path).

lock

[default FALSE] if TRUE, sockets lock once a connection has been accepted, preventing further connection attempts. This provides safety against more than one server trying to connect to a unique URL.

...

additional arguments passed through to server if launching local daemons i.e. 'url' is not specified.

monitor

(for package internal use only) do not set this parameter.

Value

Invisible NULL.

Details

The network topology is such that a dispatcher acts as a gateway between clients and servers, ensuring that tasks received from clients are dispatched on a FIFO basis to servers for processing. Tasks are queued at the dispatcher to ensure tasks are only sent to servers that can begin immediate execution of the task.