URL Constructors

host_url constructs a valid host URL (at which daemons may connect) based on the computer's hostname. This may be supplied directly to the 'url' argument of daemons.

local_url constructs a random URL suitable for local daemons.

host_url(ws = FALSE, tls = FALSE, port = 0)

local_url()

Arguments

ws

[default FALSE] logical value whether to use a WebSockets 'ws://' or else TCP 'tcp://' scheme.

tls

[default FALSE] logical value whether to use TLS in which case the scheme used will be either 'wss://' or 'tls+tcp://' accordingly.

port

[default 0] numeric port to use. This should be open to connections from the network addresses the daemons are connecting from. '0' is a wildcard value that automatically assigns a free ephemeral port.

Value

A character string comprising a valid URL.

Details

host_url relies on using the host name of the computer rather than an IP address and typically works on local networks, although this is not always guaranteed. If unsuccessful, substitute an IPv4 or IPv6 address in place of the hostname.

local_url generates a random URL for the platform's default inter-process communications transport: abstract Unix domain sockets on Linux, Unix domain sockets on MacOS, Solaris and other POSIX platforms, and named pipes on Windows.

Examples

host_url()
#> [1] "tcp://fv-az1022-934:0"
host_url(ws = TRUE)
#> [1] "ws://fv-az1022-934:0"
host_url(tls = TRUE)
#> [1] "tls+tcp://fv-az1022-934:0"
host_url(ws = TRUE, tls = TRUE, port = 5555)
#> [1] "wss://fv-az1022-934:5555"

local_url()
#> [1] "abstract://a1e18e94ce2c0444fd029f10"