Custom Serialization Functions

Registers custom serialization and unserialization functions for sending and receiving external pointer reference objects.

serialization(refhook = list())

Arguments

refhook

either a list or pairlist of two functions: the signature for the first must accept a list of external pointer type objects and return a raw vector, e.g. torch::torch_serialize, and the second must accept a raw vector and return a list of external pointer type objects, e.g. torch::torch_load,
or else NULL to reset.

Value

Invisibly, the pairlist of currently-registered 'refhook' functions. A message is printed to the console when functions are successfully registered or reset.

Details

Calling without any arguments returns the pairlist of currently-registered 'refhook' functions.

This function may be called prior to or after setting daemons, with the registered functions applying across all compute profiles.

Examples

r <- serialization(list(function(x) serialize(x, NULL), unserialize))
#> [ mirai ] serialization functions registered
print(serialization())
#> [[1]]
#> function(x) serialize(x, NULL)
#> <environment: 0x56253c243448>
#> 
#> [[2]]
#> function (connection, refhook = NULL) 
#> {
#>     if (typeof(connection) != "raw" && !is.character(connection) && 
#>         !inherits(connection, "connection")) 
#>         stop("'connection' must be a connection")
#>     .Internal(unserialize(connection, refhook))
#> }
#> <bytecode: 0x56253c2751f0>
#> <environment: namespace:base>
#> 
serialization(r)
#> [ mirai ] serialization functions registered

serialization(NULL)
#> [ mirai ] serialization functions cancelled
print(serialization())
#> [[1]]
#> NULL
#> 
#> [[2]]
#> NULL
#>