Status Information

Retrieve status information for the specified compute profile, comprising current connections and daemons status.

Usage

status(.compute = NULL)

Arguments

.compute

[default NULL] character value for the compute profile to query, or NULL to query the 'default' profile.

or a 'miraiCluster' to obtain its status.

Value

A named list comprising:

connections - integer number of active daemon connections.

daemons - character URL at which host / dispatcher is listening, or else 0L if daemons have not yet been set.

mirai (present only if using dispatcher) - a named integer vector comprising: awaiting - number of tasks queued for execution at dispatcher, executing - number of tasks sent to a daemon for execution, and completed - number of tasks for which the result has been received (either completed or cancelled).

Events

If dispatcher is used combined with daemon IDs, an additional element events will report the positive integer ID when the daemon connects and the negative value when it disconnects. Only the events since the previous status query are returned.

Examples

if (FALSE) { # interactive()
status()
daemons(url = "tcp://[::1]:0")
status()
daemons(0)
}