Status Information

Retrieve status information for the specified compute profile, comprising current connections and daemons status.

status(.compute = "default")

Arguments

.compute

[default 'default'] character compute profile (each compute profile has its own set of daemons for connecting to different resources).

or a 'miraiCluster' to obtain its status.

Value

A named list comprising:

connections - integer number of active connections.
Using dispatcher: Always 1L as there is a single connection to dispatcher, which connects to the daemons in turn.

daemons - of variable type.
Using dispatcher: a status matrix (see Status Matrix section below), or else an integer 'errorValue' if communication with dispatcher failed.
Not using dispatcher: the character host URL.
Not set: 0L.

Status Matrix

When using dispatcher, $daemons comprises an integer matrix with the following columns:

i - integer index number.

online - shows as 1 when there is an active connection, or else 0 if a daemon has yet to connect or has disconnected.

instance - increments by 1 every time there is a new connection at a URL. This counter is designed to track new daemon instances connecting after previous ones have ended (due to time-outs etc.). The count becomes negative immediately after a URL is regenerated by saisei, but increments again once a new daemon connects.

assigned - shows the cumulative number of tasks assigned to the daemon.

complete - shows the cumulative number of tasks completed by the daemon.

The dispatcher URLs are stored as row names to the matrix.

Examples

if (interactive()) {
# Only run examples in interactive R sessions

status()
daemons(n = 2L, url = "wss://[::1]:0")
status()
daemons(0)

}